+++ sdlib/voting_rights.py
@@ -13,14 +13,14 @@ VOTING_RIGHTS_EXPORT_WORKSHEET_NAME = 'Voter ID For Export'
 
 class VotingRights:
     @staticmethod
-    def setup(cur, googcreds):
+    def populate(cur, googcreds, reset=False):
         """
-        Initial setup of the voting_rights table.
-        WARNING: destroys any data present in voting_rights.
+        Load data into the voting_rights table.
 
         Args:
           cur: a database cursor
           googcreds: path to a JSON file containing Google service-account credentials
+          reset: whether to empty the table before populating it (warning!)
         """
         creds = ServiceAccountCredentials.from_json_keyfile_name(googcreds,
                                                                  ['https://spreadsheets.google.com/feeds'])
@@ -29,11 +29,19 @@ class VotingRights:
         s = ss.worksheet(VOTING_RIGHTS_EXPORT_WORKSHEET_NAME)
         all_values = s.get_all_values()
 
+        # all_values[0] is the header row, from which we draw the db column names.
+        # all_values[1:] are the data rows.
+
         header = all_values[0]
+
+        # Check that the column names are valid SQL identifiers and
+        # not some injection attack a la https://xkcd.com/327/
         if any(re.search(r'\W', colname) for colname in header):
             raise 'Illegal SQL column name found among %r' % header
 
-        cur.execute('DELETE FROM voting_rights')
+        if reset:
+            cur.execute('DELETE FROM voting_rights')
+
         sql = ('INSERT INTO voting_rights (%s) VALUES (%s)' %
                (str.join(', ', header),
                 str.join(', ', ('%s',) * len(header))))
